requirements.txt
numpy==1.24.3
opencv-python==4.6.0.66
opencv-contrib-python==4.6.0.66
pillow==10.3.0
ultralytics==8.3.28